Create Reindexing_Iteration_Sorting_nlargest.py
In this code, reindexing in Python Pandas is demonstrated, allowing changes to row and column labels. Iteration in Python Pandas is shown using items(), iterrows(), and itertuples(). Sorting in Python Pandas is illustrated using sort_values() and sort_index(). Additionally, the nlargest and nsmallest methods are used to retrieve the top n largest or smallest values based on a specified column